+++ tenant/service_onboarding.go
@@ -3,6 +3,8 @@ package tenant
 import (
 	"context"
 	"fmt"
+	"strings"
+
 	"github.com/influxdata/influxdb/v2"
 	icontext "github.com/influxdata/influxdb/v2/context"
 	"github.com/influxdata/influxdb/v2/kv"
@@ -89,8 +91,28 @@ func (s *OnboardService) OnboardUser(ctx context.Context, req *influxdb.Onboardi
 
 // onboardUser allows us to onboard new users.
 func (s *OnboardService) onboardUser(ctx context.Context, req *influxdb.OnboardingRequest, permFn func(orgID, userID influxdb.ID) []influxdb.Permission) (*influxdb.OnboardingResults, error) {
-	if req == nil || req.User == "" || req.Org == "" || req.Bucket == "" {
-		return nil, ErrOnboardInvalid
+	if req == nil {
+		return nil, &influxdb.Error{
+			Code: influxdb.EEmptyValue,
+			Msg:  "onboarding failed: no request body provided",
+		}
+	}
+
+	var missingFields []string
+	if req.User == "" {
+		missingFields = append(missingFields, "username")
+	}
+	if req.Org == "" {
+		missingFields = append(missingFields, "org")
+	}
+	if req.Bucket == "" {
+		missingFields = append(missingFields, "bucket")
+	}
+	if len(missingFields) > 0 {
+		return nil, &influxdb.Error{
+			Code: influxdb.EUnprocessableEntity,
+			Msg:  fmt.Sprintf("onboarding failed: missing required fields [%s]", strings.Join(missingFields, ",")),
+		}
 	}
 
 	result := &influxdb.OnboardingResults{}