+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Because of the version bump to `go`, the macOS build for this release requires a
 1. [21866](https://github.com/influxdata/influxdb/pull/21866): Remove incorrect optimization for group-by.
 1. [21867](https://github.com/influxdata/influxdb/pull/21867): Return an error instead of panicking when InfluxQL statement rewrites fail.
 1. [21868](https://github.com/influxdata/influxdb/pull/21868): Migrate restored KV snapshots to latest schema before using them.
+1. [21869](https://github.com/influxdata/influxdb/pull/21869): Specify which fields are missing when rejecting an incomplete onboarding request.
 
 ## v2.0.7 [2021-06-04]
 ----------------------